feat: implement user permission checks for product modifications in ProductController and ProductService

+++ src/main/java/com/samazon/application/controllers/ProductController.java
@@ -22,6 +22,7 @@ import com.samazon.application.dto.common.PagedResponse;
 import com.samazon.application.dto.products.ProductRequest;
 import com.samazon.application.dto.products.ProductResponse;
 import com.samazon.application.services.ProductService;
+import com.samazon.application.utils.AuthUtil;
 import com.samazon.application.validation.OnCreateOrUpdate;
 import com.samazon.application.validation.OnPatch;
 
@@ -32,12 +33,14 @@ import lombok.AllArgsConstructor;
 @AllArgsConstructor
 public class ProductController {
 
+    private final AuthUtil authUtil;
+
     private final ProductService productService;
 
     @PostMapping("/products")
     public ResponseEntity<ProductResponse> addProduct(
             @RequestBody @Validated(OnCreateOrUpdate.class) ProductRequest request) {
-        ProductResponse createdProductResponse = productService.addProduct(request);
+        ProductResponse createdProductResponse = productService.createProduct(authUtil.getCurrentUser(), request);
         return ResponseEntity.status(HttpStatus.CREATED).body(createdProductResponse);
     }
 
@@ -49,7 +52,7 @@ public class ProductController {
             @RequestParam(name = "size", defaultValue = AppConstants.PAGE_SIZE) Integer size,
             @RequestParam(name = "sortBy", defaultValue = AppConstants.DEFAULT_SORT_BY) String sortBy,
             @RequestParam(name = "sortOrder", defaultValue = AppConstants.SORT_ORDER) String sortOrder) {
-        PagedResponse<ProductResponse> products = productService.getProducts(categoryId, keyword, page, size, sortBy,
+        PagedResponse<ProductResponse> products = productService.getAllProducts(categoryId, keyword, page, size, sortBy,
                 sortOrder);
         return ResponseEntity.ok(products);
     }
@@ -64,6 +67,7 @@ public class ProductController {
     public ResponseEntity<ProductResponse> patchProduct(
             @PathVariable Long productId,
             @RequestBody @Validated(OnPatch.class) ProductRequest request) {
+        productService.checkModificationPermission(authUtil.getCurrentUser(), productId);
         ProductResponse patchedProductResponse = productService.patchProduct(productId, request);
         return ResponseEntity.ok(patchedProductResponse);
     }
@@ -71,6 +75,7 @@ public class ProductController {
     @PutMapping("/products/{productId}")
     public ResponseEntity<ProductResponse> updateProduct(@PathVariable Long productId,
             @RequestBody @Validated(OnCreateOrUpdate.class) ProductRequest request) {
+        productService.checkModificationPermission(authUtil.getCurrentUser(), productId);
         ProductResponse updatedProductResponse = productService.updateProduct(productId, request);
         return ResponseEntity.ok(updatedProductResponse);
     }
@@ -78,12 +83,14 @@ public class ProductController {
     @PutMapping("/products/{productId}/image")
     public ResponseEntity<ProductResponse> updateProductImage(@PathVariable Long productId,
             @RequestParam("image") MultipartFile image) throws IOException {
+        productService.checkModificationPermission(authUtil.getCurrentUser(), productId);
         ProductResponse updatedProductResponse = productService.updateProductImage(productId, image);
         return ResponseEntity.ok(updatedProductResponse);
     }
 
     @DeleteMapping("/products/{productId}")
     public ResponseEntity<Void> deleteProduct(@PathVariable Long productId) {
+        productService.checkModificationPermission(authUtil.getCurrentUser(), productId);
         productService.deleteProduct(productId);
         return ResponseEntity.noContent().build();
     }
